refactor(42-03): extract createSubmoduleWorkspace and touchSubmodule to helpers.cjs
- Add createSubmoduleWorkspace with opts param (roadmap/state/phaseDir scaffolding)
- Add touchSubmodule as shared helper
- Update workspace.test.cjs to import both from helpers.cjs
- Remove local function definitions from workspace.test.cjs
